requirements.txt
streamlit
pandas
google-generativeai
Pillow
python-dotenv
minio